Add spec for OAuth authentication
