--- app/controllers/concerns/secured_with_oauth.rb
@@ -1,13 +1,0 @@
-module SecuredWithOauth
-  extend ActiveSupport::Concern
-
-  def authenticate_user!
-    return if user_authenticated?
-
-    redirect_to '/'
-  end
-
-  def user_authenticated?
-    session && session[:userinfo].present?
-  end
-end
