+++ spec/rails_helper.rb
@@ -83,4 +83,6 @@ RSpec.configure do |config|
   # config.filter_gems_from_backtrace("gem name")
 
   config.include ActiveJob::TestHelper, type: :job
+
+  config.include_context 'Mock OAuth', include_shared: true
 end
