--- spec/support/secured_with_oauth_override.rb
@@ -1,6 +1,0 @@
-# Mock users as authenticated for tests.
-module SecuredWithOauth
-  def user_authenticated?
-    true
-  end
-end
